FTR-67: -added ReservationCalendar that deals with calendar, making it extremely reusable for future
